Going HUGE

+++ lib/frequencyprocessing.c
@@ -15,7 +15,7 @@ struct signal calculateMagSquared(struct signal real_data, struct signal imag_da
     struct signal psdx;
     psdx.length = real_data.length;
     psdx.fs = real_data.fs;
-    double scaler = psdx.fs*psdx.length*2;
+    uint64_t scaler = psdx.fs*psdx.length*2;
     printf("Size of psdx: %g\n\r", psdx.length*2);
     printf("Sample rate of psdx: %g\n\r", psdx.fs);
     printf("Scaler value: %g\n\r", scaler);
